+++ examples/example1/js/LateCtrl.js
@@ -1,5 +1,0 @@
-'use strict';
-
-angular.module('TestModule').controller('LateCtrl', ['$scope', function($scope) {
-	$scope.testValue = 'Via a lazy loaded Ctrl';
-}]);