+++ examples/example1/js/app.js
@@ -6,65 +6,71 @@ var App = angular.module('app', ['ui.router', 'oc.lazyLoad'])
 		$urlRouterProvider.otherwise("/");
 		$locationProvider.hashPrefix('!');
 
-		/*$stateProvider.state('index', {
-			url: "/", // root route
-			views: {
-				"lazyLoadView": {
-					template: '<div oc-lazy-load="{name: \'TestModule\', files: [\'js/testModule.js\'], template: \'partials/partials.html\'}">sujhdjsd</div>'
-				}
-			}
-		});*/
-
-
 		// You can also load via resolve
-		$stateProvider.state('index', {
-			url: "/", // root route
-			views: {
-				"lazyLoadView": {
-					templateUrl: 'partials/testLazyLoad.html'
+		$stateProvider
+			.state('index', {
+				url: "/", // root route
+				views: {
+					"lazyLoadView": {
+						controller: 'AppCtrl', // This view will use AppCtrl loaded below in the resolve
+						templateUrl: 'partials/main.html'
+					}
+				},
+				resolve: { // Any property in resolve should return a promise and is executed before the view is loaded
+					loadMyCtrl: ['$ocLazyLoad', function($ocLazyLoad) {
+						// you can lazy load files for an existing module
+	                     return $ocLazyLoad.load({
+	                        name: 'app',
+	                        files: ['js/AppCtrl.js']
+	                     });
+					}]
 				}
-			},
-			resolve: {
-				galaxytest: ['$ocLazyLoad', function($ocLazyLoad) {
-					/*return $ocLazyLoad.load({
-						name: 'TestModule',
-						files: ['js/testModule.js']
-					});*/
+			})
+			.state('modal', {
+				parent: 'index',
+				resolve: { // Any property in resolve should return a promise and is executed before the view is loaded
+					loadOcModal: ['$ocLazyLoad', '$injector', '$rootScope', function($ocLazyLoad, $injector, $rootScope) {
+						// Load 'oc.modal' defined in the config of the provider $ocLazyLoadProvider
+	                     return $ocLazyLoad.load('oc.modal').then(function() {
+		                     $rootScope.bootstrapLoaded = true;
+		                     // inject the lazy loaded service
+		                     var $ocModal = $injector.get("$ocModal");
+		                     $ocModal.open({
+			                     url: 'modal',
+			                     cls: 'fade-in'
+		                     });
+	                     });
+					}],
 
-                    /* Or, for more than one resource...*/
-                     return $ocLazyLoad.load([{
-                        name: 'TestModule',
-                        files: ['js/testModule.js']
-                     }, {
-                         name: 'HelloGalaxy',
-                         files: ['js/helloGalaxyModule.js']
-                     }]);
-				}],
-				// request galaxytest because LateCtrl is a controler of TestModule
-				templateTest: ['$ocLazyLoad', '$rootScope', 'galaxytest', function($ocLazyLoad, $rootScope, galaxytest) {
-					$rootScope.fileRoute = 'js/LateCtrl.js';
-					return $ocLazyLoad.loadTemplateFile(['partials/partials.html', 'partials/partials2.html']);
-				}]
-			}
-		});
-
+					// resolve the sibling state and use the service lazy loaded
+					setModalBtn: ['loadOcModal', '$rootScope', '$ocModal', function(loadOcModal, $rootScope, $ocModal) {
+						$rootScope.openModal = function() {
+							$ocModal.open({
+								url: 'modal',
+								cls: 'flip-vertical'
+							});
+						}
+					}]
+				}
+			});
 
 		// Without server side support html5 must be disabled.
 		return $locationProvider.html5Mode(false);
 	}])
 	.config(['$ocLazyLoadProvider', function($ocLazyLoadProvider) {
+		// We confiture ocLazyLoad to use the lib script.js as the async loader
 		$ocLazyLoadProvider.config({
-			modules: [
-                {
-                    name: 'HelloWorld',
-                    files: ['js/helloWorldModule.js']
-                }
-//				{
-//					name: 'TestModule',
-//					files: ['js/testModule.js'],
-//					template: 'partials/testLazyLoad.html'
-//				}
-			],
-			asyncLoader: $script
+			debug: true,
+			events: true,
+			modules: [{
+				name: 'oc.modal',
+				files: [
+					'bower_components/bootstrap/dist/css/bootstrap.css', // will use the cached version if you already loaded bootstrap with the button
+					'bower_components/ocModal/dist/css/ocModal.animations.css',
+					'bower_components/ocModal/dist/css/ocModal.light.css',
+					'bower_components/ocModal/dist/ocModal.js',
+					'partials/modal.html'
+				]
+			}]
 		});
 	}]);