+++ examples/example1/js/helloGalaxyModule.js
@@ -1,9 +1,0 @@
-'use strict';
-
-angular.module('HelloGalaxy', []).
-    config(function() {
-        console.log('HelloGalaxy::Config');
-    }).
-    factory('$galaxymessage', function() {
-        return "Hello Galaxy";
-    });