+++ examples/example1/js/helloUniverseModule.js
@@ -1,5 +1,0 @@
-'use strict';
-
-angular.module('HelloUniverse', []).factory('$universemessage', function() {
-    return "Hello Universe";
-});